bitmap fix
