+++ app/src/main/java/com/cvut/arfittingroom/model/Constants.kt
@@ -10,9 +10,9 @@ const val MASK_TEXTURE_SLOT = "mask"
 const val TRANSPARENT_CODE = 0xFF_000_000
 const val TOUCH_TO_MOVE_THRESHOLD = 20f
 const val SPAN_SLOP = 7
-
+const val PREVIEW_BITMAP_SIZE = 400
+const val BITMAP_SIZE = 1024
 const val NUM_OF_ELEMENTS_IN_ROW = 3
-
 const val MAX_LOOK_NAME_LENGTH = 50
 
 // Firebase collections paths
@@ -26,7 +26,6 @@ const val LOOKS_COLLECTION = "looks"
 const val PREVIEW_COLLECTION = "preview"
 
 // Firebase attributes
-const val LOOK_ID_ATTRIBUTE = "look_id"
 const val IS_PUBLIC_ATTRIBUTE = "is_public"
 const val TYPE_ATTRIBUTE = "type"
 const val REF_ATTRIBUTE = "ref"
@@ -34,8 +33,3 @@ const val DEFAULT_COLOR_ATTRIBUTE = "default_color"
 const val PREVIEW_IMAGE_ATTRIBUTE = "preview_image"
 const val SLOT_ATTRIBUTE = "slot"
 const val AUTHOR_ATTRIBUTE = "author"
-const val IS_ANIMATED_ATTRIBUTE = "is_animated"
-const val APPLIED_MAKEUP_ATTRIBUTE = "applied_makeup"
-const val APPLIED_MODELS_ATTRIBUTE = "applied_models"
-const val NAME_ATTRIBUTE = "name"
-const val HISTORY_ATTRIBUTE = "history"
