+++ app/src/main/java/com/cvut/arfittingroom/service/StateService.kt
@@ -4,10 +4,12 @@ import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.Matrix
 import android.util.Log
+import com.cvut.arfittingroom.model.BITMAP_SIZE
 import com.cvut.arfittingroom.model.FaceNodesInfo
 import com.cvut.arfittingroom.model.MAKEUP_SLOT
 import com.cvut.arfittingroom.model.MakeupInfo
 import com.cvut.arfittingroom.model.ModelInfo
+import com.cvut.arfittingroom.utils.BitmapUtil.combineBitmaps
 import com.cvut.arfittingroom.utils.BitmapUtil.replaceNonTransparentPixels
 import com.google.ar.core.AugmentedFace
 import com.google.ar.core.TrackingState
@@ -38,23 +40,10 @@ class StateService {
             return null
         }
 
-        val combinedBitmap = Bitmap.createBitmap(1024, 1024, Bitmap.Config.ARGB_8888)
-        val canvas = Canvas(combinedBitmap)
-
-        makeUpBitmaps.forEach {
-            val scale = (1024f / it.width).coerceAtMost(1024f / it.height)
-
-            val matrix = Matrix().apply {
-                postScale(scale, scale)
-                postTranslate((1024 - it.width * scale) / 2, (1024 - it.height * scale) / 2)
-            }
-
-            canvas.drawBitmap(it, matrix, null)
-        }
-
+        val bitmap = combineBitmaps(makeUpBitmaps, BITMAP_SIZE, BITMAP_SIZE)
         makeUpBitmaps.clear()
 
-        return combinedBitmap
+        return bitmap
     }
 
     fun clearAll() {
